Fix circular import in theme.android.ts by extracting shared defs to theme.base.ts
Metro resolves `./theme` to `./theme.android.ts` on Android, causing
theme.android.ts to circularly import itself → stack overflow. Fix by
moving the shared palette/colors/type into theme.base.ts (a suffix Metro
ignores), having theme.ts re-export from there, and importing theme.base
directly in theme.android.ts.
